Reduce code size by using more sklearn functions
